+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Command to abort a Job Run.
+
 ## [1.0.4] - 2024-10-04
 
 ## Fixed